Rework custom auditing handler
