+++ fineract-provider/src/main/java/org/apache/fineract/infrastructure/core/auditing/JpaAuditingHandlerRegistrar.java
@@ -18,11 +18,11 @@
  */
 package org.apache.fineract.infrastructure.core.auditing;
 
+import org.apache.fineract.infrastructure.auditing.CustomAuditingHandler;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
 import org.springframework.context.annotation.ImportBeanDefinitionRegistrar;
 import org.springframework.core.type.AnnotationMetadata;
-import org.springframework.data.auditing.CustomAuditingHandler;
 
 public class JpaAuditingHandlerRegistrar implements ImportBeanDefinitionRegistrar {
